Migrate rayon, rework parallel sort example
Rest of rayon examples need to be redone, though they all
now pass unit tests.

For parallel sort, the example is now just 10 words instead of 100K,
and encourages experimentation on the command line. The normal versions
of the iterator and sort are shown in comments. Since a vector is needed
for converting the u8 bytes to a String, the Vector is just created and
populated with a for() loop, instead of converting the Iterator.